+++ youtubeViewer.py
@@ -1,233 +1,138 @@
-# Imports
 from datetime import datetime
 import json
 import re
 import requests
 from bs4 import BeautifulSoup
-from hashlib import blake2b
-from alive_progress import alive_bar
 
-def scrapeTrendingPage(URL, verbose=False):
 
-    # try to get the page / URL to scrape
-    try:
-        page = requests.get(URL)
-    except:
-        # print("Error: Could not scrape URL: ", URL)
-        return
-    
-    # Initialize the BeautifulSoup object using the page text
-    soup = BeautifulSoup(bytes(page.text, encoding="utf-8"), "html.parser")
 
-    # Extract the links from the page
-    script = soup.find("script", string=re.compile('var ytInitialData'))
+"""
+----------------------------------------------------------
+Populates dictionary with a chain of youtube videos by continuously getting the first suggested video for each video 
+Use: watchedDict = getVideos(maxDepth=10, rootID="suO6X5U4rmQ", api, bar)
+----------------------------------------------------------
+Parameters:
+    maxDepth - maximum depth the function will run to
+    rootID - The initial video to start the chain on
+    api - the youtube api, used to fetch video data
+    bar - alive_progress progress bar
+Returns:
+    watchedDict - All the youtube videos seen, indexed with thier youtube ID
+    deadends - Dictionary of IDs that lead to deadends
+----------------------------------------------------------
+"""
+def getVideos(maxDepth, rootID, api, bar):
+    depth = 0 # initialize depth to 0
+    IDs = [rootID] # Initialze ID stack 
+    watchedDict = {}
+    deadends = {} # Initialize a dictionary of dead end IDs to avoid
 
-    # Extract the JSON part from the script content
-    script_content = script.string
-    start_index = script_content.find('{')
-    end_index = script_content.rfind('}') + 1
-    json_str = script_content[start_index:end_index]
-    
-    yt_data = json.loads(json_str)
+    while(depth < maxDepth):
+        # add the video data to the dictionary
+        # print(currID)
 
-    # For debugging, dumps javascript variables of page to json for easier viewing
-    # with open("yt_data/" + H + ".json", "w", encoding="utf-8") as f:
-    #     json.dump(yt_data, f, indent=4)
-
-    trending_urls = []
-
-    if 'contents' in yt_data and 'twoColumnBrowseResultsRenderer' in yt_data['contents']:
-        contents = yt_data['contents']['twoColumnBrowseResultsRenderer']['tabs'][0]['tabRenderer']['content']
-        
-        for section in contents['sectionListRenderer']['contents']:
-            if 'itemSectionRenderer' in section:
-                for item in section['itemSectionRenderer']['contents']:
-                    if 'shelfRenderer' in item:
-                        try:
-                            for video_item in item['shelfRenderer']['content']['expandedShelfContentsRenderer']['items']:
-                                    if 'videoRenderer' in video_item:
-                                        video_id = video_item['videoRenderer']['videoId']
-                                        watch_url = f"https://www.youtube.com/watch?v={video_id}"
-                                        trending_urls.append(watch_url)
-                        except KeyError as e:
-                            if(verbose): print("Could not find key with error: ", e, "... skipping iteration of loop")
-                            continue
-                            
-    return trending_urls
-
-def scrapeVideo(URL, maxdepth, videos_watched, rewrite, doSuggestedVideos, bar, verbose=False, depth=0, numAutoplay=0, numSuggested=0):
-    if(depth == maxdepth):
-        # We have reached the maximum depth so we will return
-        # print("Maximum depth reached")
-        return numAutoplay, numSuggested
-    
-    # Find the hash of the URL
-    H = blake2b(bytes(URL, encoding='utf-8')).hexdigest()
-    
-    # try to get the page / URL to scrape
-    try:
-        page = requests.get(URL)
-    except:
-        # print("Error: Could not scrape URL: ", URL)
-        return numAutoplay, numSuggested
-    
-    # Write to the log file
-    if(verbose):
-        print(f"<{URL}, {depth}, {maxdepth}>")
-    with open("crawler1.log", "a") as f:
-        # <H,URL,Download DateTime, HTTP Response Code>
-        f.write("<" + H + ", " + URL + ", " + datetime.now().strftime("%d/%m/%Y %H:%M:%S") + ", " + str(page.status_code) + ">\n")
-
-    # Initialize the BeautifulSoup object using the page text
-    soup = BeautifulSoup(bytes(page.text, encoding="utf-8"), "html.parser")
-
-    # Extract the links from the page
-    script = soup.find("script", string=re.compile('var ytInitialData'))
-
-    # Extract the JSON part from the script content
-    script_content = script.string
-    start_index = script_content.find('{')
-    end_index = script_content.rfind('}') + 1
-    json_str = script_content[start_index:end_index]
-    
-    yt_data = json.loads(json_str)
-
-    # For debugging, dumps javascript variables of page to json for easier viewing
-    # with open(f"yt_data/{depth}-" + H + ".json", "w", encoding="utf-8") as f:
-    #     json.dump(yt_data, f, indent=4)
-
-    video_data = {}
-
-    # -------------video title-------------
-    try:
-        video_data["title"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][0]['videoPrimaryInfoRenderer']['title']['runs'][0]['text']
-    except KeyError as e:
-        if(verbose): print("Could not find key:", e, "... inputing N/A")
-        video_data["title"] = "N/A"
-    # -------------video channel-------------
-    try:
-        video_data["channel"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][1]['videoSecondaryInfoRenderer']['owner']['videoOwnerRenderer']['title']['runs'][0]['text']
-    except KeyError as e:
-        if(verbose): print("Could not find key:", e, "... inputing N/A")
-        video_data["channel"] = "N/A"
-    try:
-        video_data["channel_tag"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][1]['videoSecondaryInfoRenderer']['owner']['videoOwnerRenderer']['title']['runs'][0]['navigationEndpoint']['commandMetadata']['webCommandMetadata']['url']
-    except KeyError as e:
-        if(verbose): print("Could not find key:", e, "..., inputing N/A")
-        video_data["channel_tag"] = "N/A"
-    # -------------upload date-------------
-    try:
-        video_data["upload_date"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][0]['videoPrimaryInfoRenderer']['dateText']['simpleText']
-    except KeyError as e:
-        if(verbose): print("Could not find key:", e, "..., inputing N/A")
-        video_data["upload_date"] = "N/A"
-    # -------------views-------------
-    try:
-        video_data["view_count"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][0]['videoPrimaryInfoRenderer']['viewCount']['videoViewCountRenderer']['viewCount']['simpleText']
-    except KeyError as e:
-        if(verbose): print("Could not find key:", e, "..., inputing N/A")
-        video_data["view_count"] = "N/A"
-    # -------------description-------------
-    try:
-        video_data["description"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][1]['videoSecondaryInfoRenderer']['attributedDescription']['content']
-    except KeyError as e:
-        if(verbose): print("Could not find key:", e, "... inputing N/A")
-        video_data["description"] = "N/A"
-    # -------------Likes-------------
-    try:
-        video_data["likes"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][0]['videoPrimaryInfoRenderer']['videoActions']['menuRenderer']['topLevelButtons'][0]['segmentedLikeDislikeButtonViewModel']['likeCountEntity']['likeCountIfIndifferentNumber']
-    except KeyError as e:
-        if(verbose): print("Could not find Likes key:", e, "... trying simplified like number")
+        # check if we have visited this id before and update seen count and depth list
         try:
-            video_data["likes"] = yt_data['contents']['twoColumnWatchNextResults']['results']['results']['contents'][0]['videoPrimaryInfoRenderer']['videoActions']['menuRenderer']['topLevelButtons'][0]['segmentedLikeDislikeButtonViewModel']['likeButtonViewModel']['likeButtonViewModel']['toggleButtonViewModel']['toggleButtonViewModel']['defaultButtonViewModel']['buttonViewModel']['title']
-        except KeyError as e:
-            if(verbose): print("Could not find Likes key:", e, "... entering N/A.")
-            video_data["likes"] = "N/A"
-    # -------------CURRENT URL-------------
-    video_data["URL"] = URL
-    # -------------NEXT URL-------------
-    if(doSuggestedVideos): # if doSuggestedVideos is true, get suggested video URL first
-        next_video_id, error = getSuggestedVideo(yt_data)
-        if(next_video_id != ""):
-            numSuggested += 1
-        else:
-            # if suggested video fails try autoplay
-            if(verbose): print("Could not find suggested video key:", error, "..., trying autoplay video")
-            if(error != ""):
-                with open(f"yt_data/depth-{depth}-Suggested_error-KEYERROR-{error}-" + H + ".json", "w", encoding="utf-8") as f:
-                        json.dump(yt_data, f, indent=4)
-            else:
-                with open(f"yt_data/depth-{depth}-Suggested_not_found-" + H + ".json", "w", encoding="utf-8") as f:
-                        json.dump(yt_data, f, indent=4)
+            watchedDict[IDs[-1]]["seenCount"] += 1
+            watchedDict[IDs[-1]]['depths'].append(depth)
+        except KeyError:
+            try:
+                watchedDict[IDs[-1]] = api.get_video_by_id(video_id=IDs[-1]).to_dict()
+            except:
+                print(IDs[-2], IDs[-1])
+                with open(f"errors/{IDs[-2]}.json", "w", encoding="utf-8") as f:
+                    json.dump(yt_data, f, indent=4)
+                exit(-1)
+            watchedDict[IDs[-1]]["seenCount"] = 1
+            watchedDict[IDs[-1]]['depths'] = [depth]
 
-            next_video_id, error = getAutoplayVideo(yt_data)
-            if(next_video_id != ""):
-                numAutoplay += 1
-            else:
-                # if both fail return
-                print("KEY ERROR: ", error)
-                with open(f"yt_data/depth-{depth}-Autoplay_error-KEYERROR-{error}-" + H + ".json", "w", encoding="utf-8") as f:
-                    json.dump(yt_data, f, indent=4)
-                print("Could not find the next video..., returning at depth: ", depth)
-                return numAutoplay, numSuggested
-    else: # Otherwise get next autoplay video
-        next_video_id, error = getAutoplayVideo(yt_data)
-        if(next_video_id != ""):
-            numAutoplay += 1
-        else:
-            # if autoplay video fails try suggested
-            if(verbose): print("Could not find suggested video key:", e, "..., trying autoplay video")
-            next_video_id = getSuggestedVideo(yt_data)
-            if(next_video_id != ""):
-                numSuggested += 1
-            else:
-                # if both fail return
-                print("KEY ERROR: ", error)
-                with open(f"yt_data/depth-{depth}-KEYERROR-{error}-" + H + ".json", "w", encoding="utf-8") as f:
-                    json.dump(yt_data, f, indent=4)
-                print("Could not find the next video..., returning at depth: ", depth)
-                return numAutoplay, numSuggested
-    next_watch_url = f"https://www.youtube.com/watch?v={next_video_id}"
+
+        # try to get the page / URL to scrape
+        URL = f"https://youtube.com/watch?v={IDs[-1]}" # set URL to current ID
+        try:
+            page = requests.get(URL)
+        except:
+            # print("Error: Could not scrape URL: ", URL)
+            return -1
+        # Write the page to the log file
+        with open("crawler1.log", "a") as f:
+            # <ID,URL,Download DateTime, HTTP Response Code>
+            f.write("<" + IDs[-1] + ", " + URL + ", " + datetime.now().strftime("%d/%m/%Y %H:%M:%S") + ", " + str(page.status_code) + ">\n")
+        
+        # Initialize the BeautifulSoup object using the page text
+        soup = BeautifulSoup(bytes(page.text, encoding="utf-8"), "html.parser")
+
+        # Extract the links from the page
+        script = soup.find("script", string=re.compile('var ytInitialData'))
+
+        # Extract the JSON part from the script content
+        script_content = script.string
+        start_index = script_content.find('{')
+        end_index = script_content.rfind('}') + 1
+        json_str = script_content[start_index:end_index]
     
-    # check if we have seen this video before
+        yt_data = json.loads(json_str)
+        # keep getting suggested video until
+        newID = getSuggestedVideo(yt_data, deadends)
+        if(newID == -1):
+            # could not find new ID, label this ID as a dead end 
+            print(f"{IDs[-1]} is deadend! Returning to previous ID.")
+            deadends[IDs[-1]] = IDs[-1]
+            IDs.pop()
+        else:
+            IDs.append(newID)
+            depth += 1
+            bar() # update progress bar
+    
+    return watchedDict, deadends
+
+"""
+----------------------------------------------------------
+Finds the first suggested video ID from the given json dictionary
+Use: ID = getSuggestedVideo(yt_data, deadends)
+----------------------------------------------------------
+Parameters:
+    yt_data - Dictionary of the javascript variables stored in the youtube page, this should have a suggested video and can come in different forms.
+    deadends - Dictionary of IDs that do not lead to new videos, avoid these IDs.
+Returns:
+    ID - ID of the first suggested video, -1 if video could not be found
+----------------------------------------------------------
+"""
+def getSuggestedVideo(yt_data, deadends):
+    # find the next suggested video from a given youtube page
     try:
-        videos_watched[H]["seen_count"] += 1
-        videos_watched[H]['depths'].append(depth)
+        ID = yt_data["responseContext"]["webResponseContextExtensionData"]["webPrefetchData"]["navigationEndpoints"][0]["watchEndpoint"]["videoId"]
+        return ID 
     except KeyError:
-        videos_watched[H] = video_data
-        videos_watched[H]["seen_count"] = 1
-        videos_watched[H]['depths'] = [depth]
+        ""
 
-    bar()
-    return scrapeVideo(next_watch_url, maxdepth, videos_watched, rewrite, doSuggestedVideos, bar, verbose, depth+1, numAutoplay, numSuggested)  
-
-def getSuggestedVideo(yt_data):
-    next_video_id = ""
-    
-    try: # Try catch to check if the results are there
-        for result in yt_data['contents']['twoColumnWatchNextResults']['secondaryResults']['secondaryResults']['results']:
-            try: # Try catch to check if this is the compactVideoRenderer JSON
-                next_video_id = result['compactVideoRenderer']['videoId']
-                return next_video_id, ""
-            except KeyError:
-                try: # Try catch to check if this is the compactMovieRenderer JSON
-                    next_video_id = result['compactMovieRenderer']['videoId']
-                    return next_video_id, ""
-                except KeyError:
-                    for content in result['richGridRenderer']['contents']:
-                        try: # Try catch to check if this is the richItemRenderer JSON
-                            next_video_id = content['richItemRenderer']['content']['videoRenderer']['videoId']
-                            return next_video_id, ""
-                        except KeyError as e:
-                            # Very bad, cant get video
-                            return "", e
-                continue
-    except KeyError as e:
-        return "", e
-    return next_video_id, ""
-def getAutoplayVideo(yt_data):
+    # Try compactVideoRenderer
     try:
-        next_video_id = yt_data['contents']['twoColumnWatchNextResults']['autoplay']['autoplay']['sets'][0]['autoplayVideo']['watchEndpoint']['videoId']
-    except KeyError as e:
-        return "", e
-    return next_video_id, ""
+        for result in yt_data["contents"]["twoColumnWatchNextResults"]["secondaryResults"]["secondaryResults"]["results"]:
+            ID = result["compactVideoRenderer"]["videoId"]
+            try:
+                # if ID is in deadends, continue the loop
+                deadends[ID]
+                continue
+            except KeyError:
+                # If ID is not in deadends return it.
+                return ID
+    except KeyError:
+        ""
+
+    # Try richGridRenderer
+    try:
+        for item in yt_data["contents"]["twoColumnWatchNextResults"]["secondaryResults"]["secondaryResults"]["results"][0]["richGridRenderer"]["contents"]:
+            ID = item["richItemRenderer"]["content"]["videoRenderer"]["videoId"]
+            try:
+                # if ID is in deadends, continue the loop
+                deadends[ID]
+                continue
+            except KeyError:
+                # If ID is not in deadends return it.
+                return ID
+    except KeyError:
+        ""
+    return -1
+
+
